Adding a Database and connection
Addead a database and connection so combo boxes and other dataset are loaded from the Database.
For now, the combobox for NPC Vendors ID is loaded from the Database. Currently the database is created outside, but I think we should be able to edit it from inside the  application, when more NPC Vendors are added, for example.